+++ src/main/java/com/oss/gitborad/repository/CategoryGroupRepository.java
@@ -1,4 +1,4 @@
-package com.oss.gitborad.data.repository;
+package com.oss.gitborad.repository;
 
 import com.oss.gitborad.data.domain.CategoryGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
