+++ src/main/java/com/oss/gitborad/repository/ProjectRepository.java
@@ -1,4 +1,4 @@
-package com.oss.gitborad.data.repository;
+package com.oss.gitborad.repository;
 
 import com.oss.gitborad.data.domain.Category;
 import com.oss.gitborad.data.domain.Project;
