+++ src/main/java/com/oss/gitborad/service/impl/BookmarkServiceImpl.java
@@ -3,8 +3,8 @@ package com.oss.gitborad.service.impl;
 import com.oss.gitborad.data.domain.Bookmark;
 import com.oss.gitborad.data.domain.User;
 import com.oss.gitborad.data.dto.BookmarkDTO;
-import com.oss.gitborad.data.repository.BookmarkRepository;
-import com.oss.gitborad.data.repository.ProjectRepository;
+import com.oss.gitborad.repository.BookmarkRepository;
+import com.oss.gitborad.repository.ProjectRepository;
 import com.oss.gitborad.repository.UserRepository;
 import com.oss.gitborad.service.BookmarkService;
 import org.springframework.data.domain.Page;
