+++ src/main/java/com/oss/gitborad/service/impl/CategoryServiceImpl.java
@@ -4,8 +4,8 @@ import com.oss.gitborad.data.domain.Category;
 import com.oss.gitborad.data.domain.CategoryGroup;
 import com.oss.gitborad.data.domain.User;
 import com.oss.gitborad.data.dto.CategoryDTO;
-import com.oss.gitborad.data.repository.CategoryGroupRepository;
-import com.oss.gitborad.data.repository.CategoryRepository;
+import com.oss.gitborad.repository.CategoryGroupRepository;
+import com.oss.gitborad.repository.CategoryRepository;
 import com.oss.gitborad.repository.UserRepository;
 import com.oss.gitborad.service.CategoryService;
 import org.springframework.data.domain.Page;
