+++ src/main/java/com/oss/gitborad/service/impl/InterestServiceImpl.java
@@ -4,8 +4,8 @@ import com.oss.gitborad.data.domain.Category;
 import com.oss.gitborad.data.domain.Interest;
 import com.oss.gitborad.data.domain.User;
 import com.oss.gitborad.data.dto.InterestDTO;
-import com.oss.gitborad.data.repository.CategoryRepository;
-import com.oss.gitborad.data.repository.InterestRepository;
+import com.oss.gitborad.repository.CategoryRepository;
+import com.oss.gitborad.repository.InterestRepository;
 import com.oss.gitborad.repository.UserRepository;
 import com.oss.gitborad.service.InterestService;
 import org.springframework.data.domain.Page;
