+++ src/main/java/com/oss/gitborad/service/impl/ProjectServiceImpl.java
@@ -5,9 +5,9 @@ import com.oss.gitborad.data.domain.Project;
 import com.oss.gitborad.data.domain.ProjectCategory;
 import com.oss.gitborad.data.domain.User;
 import com.oss.gitborad.data.dto.ProjectDTO;
-import com.oss.gitborad.data.repository.CategoryRepository;
-import com.oss.gitborad.data.repository.ProjectCategoryRepository;
-import com.oss.gitborad.data.repository.ProjectRepository;
+import com.oss.gitborad.repository.CategoryRepository;
+import com.oss.gitborad.repository.ProjectCategoryRepository;
+import com.oss.gitborad.repository.ProjectRepository;
 import com.oss.gitborad.repository.UserRepository;
 import com.oss.gitborad.service.ProjectService;
 import org.springframework.data.domain.Page;
