+++ src/main/java/com/oss/gitborad/service/impl/UserServiceImpl.java
@@ -3,8 +3,8 @@ package com.oss.gitborad.service.impl;
 import com.oss.gitborad.data.domain.User;
 import com.oss.gitborad.data.domain.UserBadge;
 import com.oss.gitborad.data.dto.UserDTO;
-import com.oss.gitborad.data.repository.BadgeRepository;
-import com.oss.gitborad.data.repository.UserBadgeRepository;
+import com.oss.gitborad.repository.BadgeRepository;
+import com.oss.gitborad.repository.UserBadgeRepository;
 import com.oss.gitborad.repository.UserRepository;
 import com.oss.gitborad.security.oauth.OAuthAttributes;
 import com.oss.gitborad.service.UserService;
